+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+- local data config: if the read_policy is not provided for a local data attribute, it is assumed to be a backend only attribute
+  (referencing it in a frontend will cause a 400 local data attribute undefined)
+- prepared filter config: if the use_policy is not provided for a prepared filter, it is assumed to be a backend only filter
+  (referencing it in a frontend will cause a 400 filter undefined)
+
 ## v0.1.243
 
 - AbstractDataProvider: throw 500 instead of 400 in case a filter exception occurs on creation